requirements.txt
flask
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-uploads